+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added statistics about not-yet-completed Task Assignments to
+  Project Stats page.
+
 ### Changed
 - Task Assignments from inactive Batches and/or Projects are no
   longer included in the list of "Abandoned Tasks" shown to